Fading working

+++ src/scenestorage.h
@@ -9,8 +9,8 @@ typedef struct {
     citycolor_data_t citycolor[CITYCOLOR_FIXTURE_COUNT];
 
     uint8_t scenInactive;
-    uint8_t crossfade;
-    uint8_t timing;
+    uint16_t crossfade;
+    uint16_t timing;
 } scene_t;
 
 void scenestorage_saveTo(uint8_t index);
@@ -18,8 +18,8 @@ void scenestorage_load(uint8_t index);
 
 void scenestorage_delete(uint8_t index);
 uint8_t scenestorage_hasData(uint8_t index);
-uint8_t scenestorage_getTiming(uint8_t index);
-uint8_t scenestorage_getCrossfade(uint8_t index);
-void scenestorage_setTiming(uint8_t index, uint8_t timing);
-void scenestorage_setCrossfade(uint8_t index, uint8_t crossfade);
+uint16_t scenestorage_getTiming(uint8_t index);
+uint16_t scenestorage_getCrossfade(uint8_t index);
+void scenestorage_setTiming(uint8_t index, uint16_t timing);
+void scenestorage_setCrossfade(uint8_t index, uint16_t crossfade);
 void scenestorage_getScene(uint8_t index, scene_t* scene);
